Gradle 4.10.2

+++ packr-gradle-plugin/src/com/hendraanggrian/packr/PackTask.kt
@@ -4,12 +4,13 @@ import com.badlogicgames.packr.Packr
 import com.badlogicgames.packr.PackrConfig
 import com.badlogicgames.packr.PackrConfig.Platform
 import com.hendraanggrian.packr.dist.Distribution
-import com.hendraanggrian.packr.dist.MacDistribution
+import com.hendraanggrian.packr.dist.MacOSDistribution
 import org.gradle.api.DefaultTask
+import org.gradle.api.logging.LogLevel.INFO
 import org.gradle.api.tasks.Classpath
 import org.gradle.api.tasks.Input
-import org.gradle.api.tasks.InputDirectory
 import org.gradle.api.tasks.InputFiles
+import org.gradle.api.tasks.OutputDirectory
 import org.gradle.api.tasks.TaskAction
 import java.awt.Desktop
 import java.io.File
@@ -19,25 +20,25 @@ import java.io.IOException
 open class PackTask : DefaultTask(), VMArged {
 
     companion object {
-        const val MINIMIZATION_SOFT = "soft"
-        const val MINIMIZATION_HARD = "hard"
-        const val MINIMIZATION_ORACLEJRE8 = "oraclejre8"
+        const val MINIMIZE_SOFT = "soft"
+        const val MINIMIZE_HARD = "hard"
+        const val MINIMIZE_ORACLEJRE8 = "oraclejre8"
     }
 
     private val packr = Packr()
-    private val distributions: MutableCollection<Distribution> = mutableListOf()
+    private val distributions = mutableListOf<Distribution>()
 
     /**
      * Name of the native executable, without extension such as `.exe`.
      * Default is project's name.
      */
-    @Input var executable: String? = null
+    @Input var executable: String = ""
 
     /**
      * File locations of the JAR files to package.
      * Default is empty.
      */
-    @Classpath @InputFiles var classpath: MutableCollection<String> = mutableListOf()
+    @Classpath @InputFiles var classpath: MutableList<String> = mutableListOf()
 
     /**
      * The fully qualified name of the main class, using dots to delimit package names.
@@ -49,25 +50,25 @@ open class PackTask : DefaultTask(), VMArged {
      * List of files and directories to be packaged next to the native executable.
      * Default is empty.
      */
-    @InputFiles var resources: MutableCollection<String> = mutableListOf()
+    @InputFiles var resources: MutableList<File> = mutableListOf()
 
     /**
      * Minimize the JRE by removing directories and files as specified by an additional config file.
      * Comes with a few config files out of the box.
      * Default is `soft`.
      */
-    @Input var minimization: String = MINIMIZATION_SOFT
+    @Input var minimizeJre: String = MINIMIZE_SOFT
 
     /**
      * The output directory.
      * Default is `release` directory in project's build directory.
      */
-    @InputDirectory var outputDir: String? = null
+    @OutputDirectory lateinit var outputDir: File
 
     @Input override var vmArgs: MutableCollection<String> = mutableListOf()
 
     /**
-     * Print extra messages about JRE minimization when set to `true`.
+     * Print extra messages about JRE minimizeJre when set to `true`.
      * This is an optional property.
      */
     @Input var verbose: Boolean = false
@@ -79,63 +80,73 @@ open class PackTask : DefaultTask(), VMArged {
     @Input var openOnDone: Boolean = false
 
     /** Configure macOS distribution. Unlike other distributions, mac configuration have some OS-specific properties. */
-    @JvmOverloads fun mac(init: (MacDistribution.() -> Unit)? = null) {
-        distributions += MacDistribution(project).apply { if (init != null) init() }
+    @JvmOverloads fun macOS(init: (MacOSDistribution.() -> Unit)? = null) {
+        distributions += MacOSDistribution(project.name).also { init?.invoke(it) }
     }
 
     /** Configure Windows 32-bit distribution. */
     @JvmOverloads fun windows32(init: (Distribution.() -> Unit)? = null) {
-        distributions += Distribution(project, Platform.Windows32).apply { if (init != null) init() }
+        distributions += Distribution(Platform.Windows32, project.name).also { init?.invoke(it) }
     }
 
     /** Configure Windows 64-bit distribution. */
     @JvmOverloads fun windows64(init: (Distribution.() -> Unit)? = null) {
-        distributions += Distribution(project, Platform.Windows64).apply { if (init != null) init() }
+        distributions += Distribution(Platform.Windows64, project.name).also { init?.invoke(it) }
     }
 
     /** Configure Linux 32-bit distribution. */
     @JvmOverloads fun linux32(init: (Distribution.() -> Unit)? = null) {
-        distributions += Distribution(project, Platform.Linux32).apply { if (init != null) init() }
+        distributions += Distribution(Platform.Linux32, project.name).also { init?.invoke(it) }
     }
 
     /** Configure Linux 64-bit distribution. */
     @JvmOverloads fun linux64(init: (Distribution.() -> Unit)? = null) {
-        distributions += Distribution(project, Platform.Linux64).apply { if (init != null) init() }
+        distributions += Distribution(Platform.Linux64, project.name).also { init?.invoke(it) }
     }
 
     @TaskAction
     @Throws(IOException::class)
     fun pack() {
-        check(mainClass.isNotEmpty()) { "Undefined main class" }
-        distributions.forEach { dist ->
-            println() // extra line to separate distribution processes
-            println("Packing ${dist.platform}:")
+        check(mainClass.isNotEmpty()) { "Undefined main class." }
+
+        distributions.forEach { it ->
+            logger.log(INFO, "Creating configuration for ${it.platform}.")
 
             val config = PackrConfig()
-            config.platform = dist.platform
-            config.jdk = checkNotNull(dist.jdk) { "JDK path has not yet been initialized" }
-            config.executable = executable!!
-            config.classpath = classpath.map { project.projectDir.resolve(it).path }
+            config.platform = it.platform
+            config.jdk = checkNotNull(it.jdk) { "JDK path has not yet been specified." }
+            config.executable = executable
+            config.classpath = classpath
             config.mainClass = mainClass
-            config.outDir = project.projectDir.resolve(outputDir!!).resolve(dist.name)
-            config.vmArgs = vmArgs + dist.vmArgs
-            config.resources = resources.map { project.projectDir.resolve(it) }
-            config.minimizeJre = minimization
-            if (dist is MacDistribution) {
-                if (dist.icon != null) config.iconResource = project.projectDir.resolve(dist.icon!!)
-                if (dist.bundleId != null) config.bundleIdentifier = dist.bundleId
+            config.outDir = outputDir.resolve(it.name)
+            config.vmArgs = vmArgs + it.vmArgs
+            config.resources = resources
+            config.minimizeJre = minimizeJre
+            if (it is MacOSDistribution) {
+                if (it.icon != null) config.iconResource = it.icon
+                if (it.bundleId != null) config.bundleIdentifier = it.bundleId
             }
             config.verbose = verbose
 
-            config.outDir.deleteRecursively()
+            when {
+                config.outDir.exists() -> {
+                    logger.log(INFO, "Deleting old output.")
+                    config.outDir.deleteRecursively()
+                }
+                else -> {
+                    logger.log(INFO, "Preparing output.")
+                    outputDir.mkdirs()
+                }
+            }
             packr.pack(config)
+            logger.log(INFO, "Pack completed.")
         }
 
         if (openOnDone) {
             require(Desktop.isDesktopSupported()) { "Desktop is not supported, disable `openOnDone`." }
             Desktop.getDesktop().run {
                 require(isSupported(Desktop.Action.OPEN)) { "Opening folder is not supported, disable `openOnDone`." }
-                open(File(outputDir))
+                open(outputDir)
             }
         }
     }
